+++ .gitignore
@@ -10,6 +10,7 @@ examples/cryptodata-plugin/cryptodata-plugin
 examples/cryptodata-proto-plugin/cryptodata-proto-plugin
 examples/datasync-plugin/datasync-plugin
 examples/etcd-lib/editor/editor
+examples/etcd-lib/election/election
 examples/etcd-lib/view/view
 examples/etcd-lib/watcher/watcher
 examples/flags-lib/flags-lib
