+++ exec/processmanager/plugin.go
@@ -19,9 +19,9 @@ import (
 	"os"
 	"os/exec"
 
-	"github.com/ligato/cn-infra/processmanager/status"
-	"github.com/ligato/cn-infra/processmanager/template"
-	"github.com/ligato/cn-infra/processmanager/template/model/process"
+	"github.com/ligato/cn-infra/exec/processmanager/status"
+	"github.com/ligato/cn-infra/exec/processmanager/template"
+	"github.com/ligato/cn-infra/exec/processmanager/template/model/process"
 
 	"github.com/ligato/cn-infra/infra"
 	"github.com/pkg/errors"
