+++ exec/processmanager/plugin_test.go
@@ -17,9 +17,9 @@ package processmanager_test
 import (
 	"testing"
 
-	"github.com/ligato/cn-infra/processmanager"
-	"github.com/ligato/cn-infra/processmanager/status"
-	tmpModel "github.com/ligato/cn-infra/processmanager/template/model/process"
+	"github.com/ligato/cn-infra/exec/processmanager"
+	"github.com/ligato/cn-infra/exec/processmanager/status"
+	tmpModel "github.com/ligato/cn-infra/exec/processmanager/template/model/process"
 	. "github.com/onsi/gomega"
 )
