+++ exec/processmanager/process_impl.go
@@ -22,7 +22,7 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/ligato/cn-infra/processmanager/status"
+	"github.com/ligato/cn-infra/exec/processmanager/status"
 	"github.com/pkg/errors"
 )
